overhauled the plugin to match certain criteria for an event

+++ src/main/java/de/amin/bingo/BingoPlugin.java
@@ -10,16 +10,17 @@ import de.amin.bingo.team.TeamManager;
 import de.amin.bingo.utils.Config;
 import de.amin.bingo.utils.Localization;
 import de.amin.bingo.team.TeamGuiListener;
+import jdk.nashorn.internal.parser.JSONParser;
 import org.apache.commons.io.FileUtils;
-import org.bukkit.Bukkit;
-import org.bukkit.ChatColor;
-import org.bukkit.GameRule;
+import org.bukkit.*;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.bukkit.scoreboard.DisplaySlot;
 
+
 import java.io.File;
 import java.io.FileInputStream;
+import java.io.FileReader;
 import java.io.IOException;
 import java.util.Properties;
 
@@ -27,12 +28,15 @@ public final class BingoPlugin extends JavaPlugin {
 
     public static BingoPlugin INSTANCE;
 
+
     @Override
     public void onLoad() {
         INSTANCE = this;
+        /*
         if(Config.WORLD_RESET) {
             resetWorld();
         }
+         */
     }
 
     @Override
@@ -65,7 +69,6 @@ public final class BingoPlugin extends JavaPlugin {
         registerListeners(getServer().getPluginManager(), gameStateManager, teamManager, game);
         registerCommands(gameStateManager, game, renderer, teamManager);
 
-        Bukkit.getWorlds().forEach(world -> world.setGameRule(GameRule.KEEP_INVENTORY, true));
     }
 
     @Override
@@ -85,22 +88,28 @@ public final class BingoPlugin extends JavaPlugin {
 
     private void registerCommands(GameStateManager gameStateManager, BingoGame game, BoardRenderer renderer, TeamManager teamManager) {
         getCommand("forcestart").setExecutor(new ForceStart(this, gameStateManager));
+        getCommand("newgame").setExecutor(new ReStartBingo(gameStateManager));
         getCommand("board").setExecutor(new BoardCommand(game, gameStateManager, teamManager));
         getCommand("reroll").setExecutor(new RerollCommand(this, game, renderer, gameStateManager));
         getCommand("backpack").setExecutor(new BackPackCommand(teamManager, gameStateManager));
         getCommand("teamchat").setExecutor(new TeamChatCommand(teamManager));
     }
 
-    private void resetWorld() {
-        File propertiesFile = new File(Bukkit.getWorldContainer(), "server.properties");
-        try (FileInputStream stream = new FileInputStream(propertiesFile)) {
-            Properties properties = new Properties();
-            properties.load(stream);
+    public void resetWorld() {
+        try  {
+            for (World world : getServer().getWorlds()) {
+                if(world.getName().equals("world_bingo")) {
+                    getServer().unloadWorld(world,false);
+                }
+            }
 
-            // Getting and deleting the main world
-            File world = new File(Bukkit.getWorldContainer(), properties.getProperty("level-name"));
-            FileUtils.deleteDirectory(world);
+            File world = new File("world_bingo");
+            if (world.exists()) {
+                FileUtils.deleteDirectory(world);
+            }
 
+
+            /*
             // Creating needed directories
             world.mkdirs();
             new File(world, "data").mkdirs();
@@ -110,6 +119,12 @@ public final class BingoPlugin extends JavaPlugin {
             new File(world, "poi").mkdirs();
             new File(world, "region").mkdirs();
 
+             */
+
+            WorldCreator wg = new WorldCreator("world_bingo");
+
+            World nw = getServer().createWorld(wg);
+            nw.setGameRule(GameRule.KEEP_INVENTORY, true);
         } catch (IOException e) {
             e.printStackTrace();
         }
